ajustar rips_departments

+++ database/migrations/2025_05_11_063346_create_rips_departments_table.php
@@ -7,16 +7,19 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up(): void
     {
-        Schema::create('rips_departments', function (Blueprint $table) {
-            $table->id();
-            $table->string('code')->unique();
-            $table->string('name');
-            $table->foreignId('rips_country_id')
-                ->constrained('rips_countries')
-                ->onDelete('cascade');
+        if (!Schema::hasTable('rips_departments')) {
 
-            $table->timestamps();
-        });
+            Schema::create('rips_departments', function (Blueprint $table) {
+                $table->id();
+                $table->string('code')->unique();
+                $table->string('name');
+                $table->foreignId('rips_country_id')
+                    ->constrained('rips_countries')
+                    ->onDelete('cascade');
+
+                $table->timestamps();
+            });
+        } 
         
     }
